reply mid changed

+++ src/main/java/com/noteCup/reply/service/ReplyMM.java
@@ -2,12 +2,15 @@ package com.noteCup.reply.service;
 
 import java.util.List;
 import java.util.Locale;
+import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.MessageSource;
 import org.springframework.stereotype.Service;
 
 import com.noteCup.contents.model.domain.ContentWrapper;
+import com.noteCup.member.model.domain.MemberInfo;
+import com.noteCup.member.repository.IMemberRepository;
 import com.noteCup.reply.model.domain.Reply;
 import com.noteCup.reply.model.dto.ReplyInput;
 import com.noteCup.reply.model.vo.ReplyVO;
@@ -24,7 +27,7 @@ import com.noteCup.reply.repository.IReplyRepository;
  * --------------------------------------
  * @author	: daniel
  * @created	: 2021. 4. 29.
- * @version	: 1.0
+ * @version	: 1.0.1
  * @formatter:on
  */
 @Service
@@ -32,6 +35,9 @@ public class ReplyMM {
 	
 	@Autowired
 	IReplyRepository replyRepository;
+	
+	@Autowired
+	IMemberRepository memberRepository;
 	
 	@Autowired
 	MessageSource messages;
@@ -46,12 +52,16 @@ public class ReplyMM {
 		return replyList; 
 	}
 
-	public void create(ReplyInput replyInput) {
+	public void create(ReplyInput replyInput, String email) {
+		
+		Optional<MemberInfo> member = memberRepository.findByEmail(email);// 이 부분이 중요함...
+		
 		ContentWrapper contentWrapper = ContentWrapper.builder()
 				.cid(replyInput.getCid())
 				.build();
 		
-		Reply reply = Reply.builder().mid(replyInput.getMid())
+		
+		Reply reply = Reply.builder().mid( member.get().getMid() ) // 이 부분이 중요함...
 						.rtext(replyInput.getRtext())
 						.contentWrapper(contentWrapper).build();
 		
@@ -61,12 +71,15 @@ public class ReplyMM {
 		}
 	}
 
-	public void update(ReplyInput replyInput) {
+	public void update(ReplyInput replyInput, String email) {
+		
+		Optional<MemberInfo> member = memberRepository.findByEmail(email);
+		
 		ContentWrapper contentWrapper = ContentWrapper.builder()
 				.cid(replyInput.getCid())
 				.build();
 		
-		Reply reply = Reply.builder().mid(replyInput.getMid())
+		Reply reply = Reply.builder().mid( member.get().getMid() )
 						.rtext(replyInput.getRtext())
 						.contentWrapper(contentWrapper).build();
 		
@@ -76,8 +89,18 @@ public class ReplyMM {
 		}
 	}
 
-	public void delete(ReplyInput replyInput) {
-		replyRepository.deleteById(replyInput.getRid());//댓글 primary 번호로 지움?..
+	public void delete(ReplyInput replyInput, String email) {
+		
+		Optional<Reply> reply = replyRepository.findById(replyInput.getRid());
+		Optional<MemberInfo> member = memberRepository.findByEmail(email);
+		Long loginedId = member.get().getMid();
+		Long replyMid = reply.get().getMid();
+		
+		if(loginedId == replyMid) {
+			replyRepository.deleteById(replyInput.getRid());//댓글 primary 번호로 지움?..
+		}else {
+			messages.getMessage("reply.message.delete.fail", null, Locale.KOREAN);
+		}
 	}
